+++ lib/components/z-collection.dart
@@ -28,8 +28,10 @@ class ZCollectionState extends State<ZCollection> {
 
   @override
   void initState() {
-    super.initState();
+    buscarValorPadrao(widget.lista);
     _itemSelecionado = new ZCollectionItem();
+
+    super.initState();
   }
 
   @override
@@ -63,7 +65,7 @@ class ZCollectionState extends State<ZCollection> {
                           ),
                         )
                       : new Text(
-                          "${_itemSelecionado?.valor ?? ""}",
+                    (_itemSelecionado?.valor.length >16)? "${_itemSelecionado?.valor.substring(0, 16) ?? ""}...":_itemSelecionado?.valor,
                           style: new TextStyle(
                             color: Colors.grey,
                           ),