+++ lib/components/z-perfil.dart
@@ -4,7 +4,6 @@ import 'package:z_components/components/z-header.dart';
 import 'package:z_components/components/z-item-tile.dart';
 import 'package:z_components/components/z-perfil-item.dart';
 import 'package:z_components/components/z-text.dart';
-import 'package:z_components/components/z_navigationbar.dart';
 import 'package:z_components/config/z-tipo-header.dart';
 import 'package:z_components/config/z-tipo-textos.dart';